lib_toolchain/libmach: fix a chain of compile bugs blocking libmach.a
libmach.a itself failed to build (blocking utilities/kernel/ksize.c,
and likely acid/ar too), found while surveying utilities/ beyond
files/ for what's next. Fixed each error as it surfaced, one at a
time:

- sym.c, executable.c, atomic.h: used `u32int`/`u64int`/`u16int` (the
  old Plan9 basic-type naming) instead of `uint32`/`uint64`/`uint16`
  -- include/core/types.h's own comment already documents this
  project moved away from the u?intNN convention, these three files
  just never got migrated. include/arch/{amd64,arm64}/ureg.h had the
  same issue in their register-dump structs.
- bootexec.h: `struct i386exec` had every one of its own fields
  already commented out "UNUSED" by the original author, leaving an
  empty struct body this compiler can't parse (empty structs aren't
  legal in this C89-ish dialect). Confirmed unreferenced anywhere else
  in libmach and deleted, rather than kept as a dead stub.
- linux.c: five stub functions (ctlproc, proctextfile, ...) end with
  a sysfatal() call and no return -- correct at runtime (sysfatal
  never returns) but this compiler still requires an explicit return
  at the end of every non-void function. Added unreachable ones.
- obj.c: atol() called with a `byte*` (unsigned char) where it wants
  `char*` -- explicit cast, same-width types differing only in
  signedness.
- 5db.c, 8db.c: twelve get4() call sites already had a cast attempting
  to fix this same class of signedness mismatch, but to the wrong
  type (`long*` instead of `uint32*`, get4's real 3rd-arg type) --
  this compiler still flagged them as fatal "argument prototype
  mismatch" errors despite the existing cast. Fixed to `uint32*`.

Verified: libmach.a now builds cleanly for arm64, and
utilities/kernel/ksize.exe links and runs against it (crackhdr()
itself not recognizing this project's own ELF binaries turned out to
be a separate, deeper functional bug -- not chased further here).
Full `mk test` (hello_libc + top-level, both objtype=arm64 and the
default objtype=boot-gcc) passes.

+++ lib_toolchain/libmach/8db.c
@@ -3,6 +3,13 @@
 #include <bio.h>
 #include "mach.h"
 
+// claude: get4()'s 3rd arg is uint32* (include/arch/mach.h); every
+// call site below originally cast to `(long*)` instead, which this
+// compiler still flags as an "argument prototype mismatch" (a real,
+// fatal error here, not a warning) since long/uint32 differ in
+// signedness -- fixed to `(uint32*)`, same as 5db.c's own identical
+// gap.
+
 /*
  * i386-specific debugger interface
  */
@@ -119,7 +126,7 @@ i386trace(Map *map, uvlong pc, uvlong sp, uvlong link, Tracer trace)
 			sp += f.value-mach->szaddr;
 		}
 
-		if (get4(map, sp, (long *) &pc) < 0)
+		if (get4(map, sp, (uint32*)&pc) < 0)
 			break;
 
 		if(pc == 0)
@@ -153,7 +160,7 @@ i386frame(Map *map, uvlong addr, uvlong pc, uvlong sp, uvlong link)
 		if (s.value == addr)
 			return sp;
 
-		if (get4(map, sp, (long *)&pc) < 0)
+		if (get4(map, sp, (uint32*)&pc) < 0)
 			break;
 		sp += mach->szaddr;
 	}
@@ -1795,9 +1802,9 @@ i386foll(Map *map, uvlong pc, Rgetter rget, uvlong *foll)
 	case RET:		/* RETURN or LEAVE */
 	case Iw:		/* RETURN */
 		if (strcmp(op->proto, "LEAVE") == 0) {
-			if (get4(map, (*rget)(map, "BP"), (long*)&l) < 0)
+			if (get4(map, (*rget)(map, "BP"), (uint32*)&l) < 0)
 				return -1;
-		} else if (get4(map, (*rget)(map, mach->sp), (long*)&l) < 0)
+		} else if (get4(map, (*rget)(map, mach->sp), (uint32*)&l) < 0)
 			return -1;
 		foll[0] = l;
 		return 1;
@@ -1818,12 +1825,12 @@ i386foll(Map *map, uvlong pc, Rgetter rget, uvlong *foll)
 			/* calculate the effective address */
 		addr = i.disp;
 		if (i.base >= 0) {
-			if (get4(map, (*rget)(map, reg[i.base]), (long*)&l) < 0)
+			if (get4(map, (*rget)(map, reg[i.base]), (uint32*)&l) < 0)
 				return -1;
 			addr += l;
 		}
 		if (i.index >= 0) {
-			if (get4(map, (*rget)(map, reg[i.index]), (long*)&l) < 0)
+			if (get4(map, (*rget)(map, reg[i.index]), (uint32*)&l) < 0)
 				return -1;
 			addr += l*(1<<i.ss);
 		}
@@ -1833,7 +1840,7 @@ i386foll(Map *map, uvlong pc, Rgetter rget, uvlong *foll)
 		foll[0] = s<<4;
 		addr += 2;
 		if (i.asize == 'L') {
-			if (get4(map, addr, (long*)&l) < 0)	/* disp32 */
+			if (get4(map, addr, (uint32*)&l) < 0)	/* disp32 */
 				return -1;
 			foll[0] += l;
 		} else {					/* disp16 */
